Разделение кода на модули, m5-t1
